+++ src/commands/policy/remove.ts
@@ -1,10 +1,19 @@
-import { PluginType } from '@cloudgraph/sdk'
-import { isEmpty } from 'lodash'
-import { Flags as flags } from '@oclif/core'
-import Command from '../base'
+import { ConfiguredPlugin, PluginType } from '@cloudgraph/sdk'
+import isEmpty from 'lodash/isEmpty'
 
-export default class Remove extends Command {
+import OperationBaseCommand from '../operation'
+
+export default class RemovePolicy extends OperationBaseCommand {
   static description = 'Remove currently installed policy pack'
+
+  static aliases = [
+    'remove:policy',
+    'policy:remove',
+    'policy:rm',
+    'del:policy',
+    'rm:policy',
+    'del:policy',
+  ]
 
   static examples = [
     '$ cg policy delete',
@@ -16,46 +25,37 @@ export default class Remove extends Command {
 
   static hidden = false
 
-  static flags = {
-    'no-save': flags.boolean({
-      default: false,
-      description: 'Set to not alter lock file, just delete plugin',
-    }),
-    ...Command.flags,
-  }
-
-  static args = Command.args
-
   async run(): Promise<void> {
-    const {
-      argv,
-      flags: { 'no-save': noSave },
-    } = await this.parse(Remove)
-    const allPolicyPacks = argv
-    const manager = await this.getPluginManager(PluginType.PolicyPack)
-    const lockFile = manager.getLockFile()
-    if (isEmpty(lockFile?.policyPack)) {
-      this.logger.info('No policy packs found, have you installed any?')
-      this.exit()
-    }
-    for (const key of allPolicyPacks) {
-      await manager.removePlugin(key)
+    try {
+      const {
+        manager,
+        noSave = false,
+        plugins: pluginsRemoved = [],
+      } = await this.remove(PluginType.PolicyPack)
 
-      if (!noSave) {
-        manager.removeFromLockFile(key)
+      for (const key of pluginsRemoved) {
+        if (manager && !noSave) {
+          manager.removeFromLockFile(key)
 
-        const [provider] = key.split('-')
-        const config = this.getCGConfig()
+          const config = this.getCGConfig()
+          if (config) {
+            const configuredPolicies =
+              config.cloudGraph.plugins[PluginType.PolicyPack] || []
 
-        if (config[provider]) {
-          config[provider].policies = [
-            ...config[provider].policies.filter(
-              (policy: string) => policy !== key
-            ),
-          ]
-          this.saveCloudGraphConfigFile(config)
+            if (!isEmpty(configuredPolicies)) {
+              // Remove policy from Policy Pack Plugin array
+              config.cloudGraph.plugins[PluginType.PolicyPack] =
+                configuredPolicies.filter(
+                  (p: ConfiguredPlugin) => p.name !== key
+                )
+
+              this.saveCloudGraphConfigFile(config)
+            }
+          }
         }
       }
+    } catch (error) {
+      this.logger.stopSpinner()
     }
   }
 }